Moved dark mode switch to its on component
