+++ src/app/app.component.scss
@@ -1,5 +1,3 @@
-$active-color: #2196f3;
-
 .header {
   display: flex;
   align-items: center;
@@ -22,49 +20,4 @@ $active-color: #2196f3;
 
     margin-inline-end: 0.5rem;
   }
-}
-
-.toggle {
-  display: flex;
-  align-items: center;
-  margin-inline-start: 0.5rem;
-}
-
-.switch {
-  position: relative;
-  display: inline-block;
-  width: 60px;
-  height: 34px;
-}
-
-.slider {
-  position: absolute;
-  cursor: pointer;
-  border-radius: 34px;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: #ccc;
-  transition: 0.4s;
-
-  &::before {
-    position: absolute;
-    content: "";
-    border-radius: 50%;
-    height: 26px;
-    width: 26px;
-    left: 4px;
-    bottom: 4px;
-    background-color: #fff;
-    transition: 0.4s;
-  }
-}
-
-input:checked + .slider {
-  background-color: $active-color;
-}
-
-input:checked + .slider:before {
-  transform: translateX(26px);
 }
