+++ src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnChanges, OnInit, Renderer2 } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Image } from 'dist/ngx-darkbox-gallery/lib/model/image';
 
 @Component({
@@ -9,14 +9,6 @@ import { Image } from 'dist/ngx-darkbox-gallery/lib/model/image';
 export class AppComponent implements OnInit {
   images: Image[] = [];
 
-  darkModeEnabled: boolean;
-
-  constructor(private renderer: Renderer2) {
-    const prefersDarkScheme = window.matchMedia("(prefers-color-scheme: dark)");
-    // Call the initialization with the OS default
-    this.onDarkModeChanged(prefersDarkScheme.matches);
-  }
-
   ngOnInit(): void {
     for (let i = 1; i <= 100; i++) {
       this.images.push({
@@ -25,19 +17,6 @@ export class AppComponent implements OnInit {
         caption: 'Image number ' + i + ' from https://picsum.photos/',
         altText: 'Random image number ' + i
       });
-    }
-  }
-
-  /**
-   * Function which enables/disables the dark mode
-   * @param darkModeEnabled Whether the dark mode is enabled or not
-   */
-  onDarkModeChanged(darkModeEnabled: boolean): void {
-    this.darkModeEnabled = darkModeEnabled;
-    if (darkModeEnabled) {
-      this.renderer.addClass(document.body, 'dark-theme');
-    } else {
-      this.renderer.removeClass(document.body, 'dark-theme');
     }
   }
 }
