+++ src/app/app.module.ts
@@ -4,10 +4,12 @@ import { NgModule } from '@angular/core';
 import { AppComponent } from './app.component';
 import { NgxDarkboxGalleryModule } from '@failed-successfully/ngx-darkbox-gallery';
 import { FormsModule } from '@angular/forms';
+import { DarkModeSwitchComponent } from './dark-mode-switch/dark-mode-switch.component';
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    DarkModeSwitchComponent
   ],
   imports: [
     BrowserModule,
