+++ src/app/dark-mode-switch/dark-mode-switch.component.scss
@@ -1,0 +1,48 @@
+$active-color: #2196f3;
+
+.toggle {
+  display: flex;
+  align-items: center;
+  margin-inline-start: 0.5rem;
+}
+
+.switch {
+  position: relative;
+  display: inline-block;
+  margin-inline-start: 0.2rem;
+  margin-inline-end: 0.2rem;
+  width: 60px;
+  height: 34px;
+}
+
+.slider {
+  position: absolute;
+  cursor: pointer;
+  border-radius: 34px;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: #ccc;
+  transition: 0.4s;
+
+  &::before {
+    position: absolute;
+    content: "";
+    border-radius: 50%;
+    height: 26px;
+    width: 26px;
+    left: 4px;
+    bottom: 4px;
+    background-color: #fff;
+    transition: 0.4s;
+  }
+}
+
+input:checked + .slider {
+  background-color: $active-color;
+}
+
+input:checked + .slider:before {
+  transform: translateX(26px);
+}
